docs: ran benchmarks again
